requirements.txt
python-dotenv
requests
loguru
fire